# api/openapi.bundled.json
{
  "components": {
    "parameters": {
      "IllustIdPath": {
        "description": "Pixiv illustration ID.",
        "in": "path",
        "name": "id",
        "required": true,
        "schema": {
          "format": "int64",
          "minimum": 1,
          "type": "integer"
        }
      },
      "IllustTypeQuery": {
        "description": "Filter by illustration type. If omitted, both are returned where supported.",
        "in": "query",
        "name": "type",
        "schema": {
          "$ref": "#/components/schemas/IllustType"
        }
      },
      "MaxBookmarkIdQuery": {
        "description": "Cursor for paging through bookmarks (from `next_max_bookmark_id` in the previous response).",
        "in": "query",
        "name": "max_bookmark_id",
        "schema": {
          "format": "int64",
          "minimum": 1,
          "type": "integer"
        }
      },
      "OffsetQuery": {
        "description": "Page offset for list endpoints that use offset-based pagination.",
        "in": "query",
        "name": "offset",
        "schema": {
          "default": 0,
          "format": "int64",
          "minimum": 0,
          "type": "integer"
        }
      },
      "RestrictQuery": {
        "in": "query",
        "name": "restrict",
        "schema": {
          "$ref": "#/components/schemas/Restrict"
        }
      },
      "UserIdPath": {
        "description": "Pixiv user ID.",
        "in": "path",
        "name": "id",
        "required": true,
        "schema": {
          "format": "int64",
          "minimum": 1,
          "type": "integer"
        }
      }
    },
    "responses": {
      "BadRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Invalid parameters."
      },
      "NoContent": {
        "description": "Action succeeded; no body."
      },
      "NotFound": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Resource not found on Pixiv."
      },
      "Unauthenticated": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "No valid Pixiv session. Call `POST /auth/login` first."
      },
      "Upstream": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Pixiv upstream returned an unexpected error."
      }
    },
    "schemas": {
      "AuthStatus": {
        "properties": {
          "authenticated": {
            "type": "boolean"
          },
          "expires_at": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "user_id": {
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "user_name": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "authenticated"
        ],
        "type": "object"
      },
      "BookmarkRequest": {
        "properties": {
          "restrict": {
            "$ref": "#/components/schemas/Restrict"
          }
        },
        "type": "object"
      },
      "Error": {
        "properties": {
          "code": {
            "description": "Machine-readable error code.",
            "example": "unauthenticated",
            "type": "string"
          },
          "detail": {
            "description": "Optional additional context (e.g., upstream body).",
            "type": "string"
          },
          "message": {
            "description": "Human-readable error message.",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "FollowRequest": {
        "properties": {
          "restrict": {
            "$ref": "#/components/schemas/Restrict"
          }
        },
        "type": "object"
      },
      "HealthStatus": {
        "properties": {
          "status": {
            "example": "ok",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
      "Illust": {
        "description": "Pixiv illustration metadata. Shape mirrors pixivgo.IllustrationInfo.",
        "type": "object",
        "x-go-type": "pixivgo.IllustrationInfo",
        "x-go-type-import": {
          "name": "pixivgo",
          "path": "github.com/txperl/pixivgo"
        }
      },
      "IllustDetailResponse": {
        "description": "Wrapper for a single illust detail. Mirrors pixivgo.IllustDetailResponse.",
        "type": "object",
        "x-go-type": "pixivgo.IllustDetailResponse",
        "x-go-type-import": {
          "name": "pixivgo",
          "path": "github.com/txperl/pixivgo"
        }
      },
      "IllustPage": {
        "properties": {
          "illusts": {
            "items": {
              "$ref": "#/components/schemas/Illust"
            },
            "type": "array"
          },
          "next_max_bookmark_id": {
            "description": "Set only by `/users/{id}/bookmarks`; pass back as `max_bookmark_id`.",
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "next_offset": {
            "description": "Next offset to pass back, or null if no further page.",
            "format": "int64",
            "nullable": true,
            "type": "integer"
          }
        },
        "required": [
          "illusts"
        ],
        "type": "object"
      },
      "IllustType": {
        "enum": [
          "illust",
          "manga"
        ],
        "type": "string"
      },
      "LoginRequest": {
        "properties": {
          "refresh_token": {
            "description": "Long-lived Pixiv OAuth refresh token. See README for how to obtain.",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "type": "object"
      },
      "RankingMode": {
        "enum": [
          "day",
          "week",
          "month",
          "day_male",
          "day_female",
          "week_original",
          "week_rookie",
          "day_manga",
          "day_r18",
          "day_male_r18",
          "day_female_r18",
          "week_r18",
          "week_r18g"
        ],
        "type": "string"
      },
      "Restrict": {
        "default": "public",
        "enum": [
          "public",
          "private"
        ],
        "type": "string"
      },
      "SearchSort": {
        "default": "date_desc",
        "enum": [
          "date_desc",
          "date_asc",
          "popular_desc"
        ],
        "type": "string"
      },
      "SearchTarget": {
        "default": "partial_match_for_tags",
        "enum": [
          "partial_match_for_tags",
          "exact_match_for_tags",
          "title_and_caption",
          "keyword"
        ],
        "type": "string"
      },
      "UgoiraMetadataResponse": {
        "description": "Ugoira (animated illust) metadata. Mirrors pixivgo.UgoiraMetadataResponse.",
        "type": "object",
        "x-go-type": "pixivgo.UgoiraMetadataResponse",
        "x-go-type-import": {
          "name": "pixivgo",
          "path": "github.com/txperl/pixivgo"
        }
      },
      "User": {
        "description": "Pixiv user summary. Shape mirrors pixivgo.UserInfo.",
        "type": "object",
        "x-go-type": "pixivgo.UserInfo",
        "x-go-type-import": {
          "name": "pixivgo",
          "path": "github.com/txperl/pixivgo"
        }
      },
      "UserIllustsPage": {
        "properties": {
          "illusts": {
            "items": {
              "$ref": "#/components/schemas/Illust"
            },
            "type": "array"
          },
          "next_offset": {
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "user",
          "illusts"
        ],
        "type": "object"
      },
      "UserPreview": {
        "description": "User + a few sample illusts. Shape mirrors pixivgo.UserPreview.",
        "type": "object",
        "x-go-type": "pixivgo.UserPreview",
        "x-go-type-import": {
          "name": "pixivgo",
          "path": "github.com/txperl/pixivgo"
        }
      },
      "UserPreviewPage": {
        "properties": {
          "next_offset": {
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "user_previews": {
            "items": {
              "$ref": "#/components/schemas/UserPreview"
            },
            "type": "array"
          }
        },
        "required": [
          "user_previews"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "description": "PixivBiu v3 backend API. Read-only browsing + auth + bookmark/follow.\nPath items are split across files under `paths/` by domain;\nschemas and shared parameters stay centralised here.\n",
    "title": "PixivBiu v3 API",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthStatus"
                }
              }
            },
            "description": "Authenticated. The access token is stored server-side."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "502": {
            "$ref": "#/components/responses/Upstream"
          }
        },
        "summary": "Exchange a refresh token for an access token (and persist it).",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logout",
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          }
        },
        "summary": "Forget the current tokens and clear the state file.",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/status": {
      "get": {
        "operationId": "getAuthStatus",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthStatus"
                }
              }
            },
            "description": "Current auth status."
          }
        },
        "summary": "Report whether the server currently holds a valid access token.",
        "tags": [
          "auth"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "getHealth",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthStatus"
                }
              }
            },
            "description": "Service is healthy."
          }
        },
        "summary": "Health check",
        "tags": [
          "system"
        ]
      }
    },
    "/illusts/following": {
      "get": {
        "operationId": "listFollowingIllusts",
        "parameters": [
          {
            "$ref": "#/components/parameters/RestrictQuery"
          },
          {
            "$ref": "#/components/parameters/OffsetQuery"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IllustPage"
                }
              }
            },
            "description": "Following feed page."
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "502": {
            "$ref": "#/components/responses/Upstream"
          }
        },
        "summary": "New illusts from users the authenticated user follows.",
        "tags": [
          "illusts"
        ]
      }
    },
    "/illusts/ranking": {
      "get": {
        "operationId": "listRanking",
        "parameters": [
          {
            "in": "query",
            "name": "mode",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/RankingMode"
                }
              ],
              "default": "day"
            }
          },
          {
            "description": "Ranking date in `YYYY-MM-DD`. Omit for the latest.",
            "in": "query",
            "name": "date",
            "schema": {
              "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}$",
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/OffsetQuery"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IllustPage"
                }
              }
            },
            "description": "Ranking page."
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "502": {
            "$ref": "#/components/responses/Upstream"
          }
        },
        "summary": "Fetch the Pixiv ranking.",
        "tags": [
          "illusts"
        ]
      }
    },
    "/illusts/recommended": {
      "get": {
        "operationId": "listRecommended",
        "parameters": [
          {
            "$ref": "#/components/parameters/IllustTypeQuery"
          },
          {
            "$ref": "#/components/parameters/OffsetQuery"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IllustPage"
                }
              }
            },
            "description": "Recommendation page."
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "502": {
            "$ref": "#/components/responses/Upstream"
          }
        },
        "summary": "Recommended illusts for the authenticated user.",
        "tags": [
          "illusts"
        ]
      }
    },
    "/illusts/{id}": {
      "get": {
        "operationId": "getIllust",
        "parameters": [
          {
            "$ref": "#/components/parameters/IllustIdPath"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IllustDetailResponse"
                }
              }
            },
            "description": "Illustration detail."
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "502": {
            "$ref": "#/components/responses/Upstream"
          }
        },
        "summary": "Fetch detail for a single illustration.",
        "tags": [
          "illusts"
        ]
      }
    },
    "/illusts/{id}/bookmark": {
      "delete": {
        "operationId": "deleteBookmark",
        "parameters": [
          {
            "$ref": "#/components/parameters/IllustIdPath"
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "502": {
            "$ref": "#/components/responses/Upstream"
          }
        },
        "summary": "Remove the illustration from bookmarks.",
        "tags": [
          "illusts"
        ]
      },
      "put": {
        "operationId": "addBookmark",
        "parameters": [
          {
            "$ref": "#/components/parameters/IllustIdPath"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "502": {
            "$ref": "#/components/responses/Upstream"
          }
        },
        "summary": "Add the illustration to the authenticated user's bookmarks.",
        "tags": [
          "illusts"
        ]
      }
    },
    "/illusts/{id}/ugoira": {
      "get": {
        "operationId": "getUgoiraMetadata",
        "parameters": [
          {
            "$ref": "#/components/parameters/IllustIdPath"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UgoiraMetadataResponse"
                }
              }
            },
            "description": "Ugoira metadata."
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "502": {
            "$ref": "#/components/responses/Upstream"
          }
        },
        "summary": "Frame timing + zip URL for an animated illustration.",
        "tags": [
          "illusts"
        ]
      }
    },
    "/search/illusts": {
      "get": {
        "operationId": "searchIllusts",
        "parameters": [
          {
            "in": "query",
            "name": "word",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search_target",
            "schema": {
              "$ref": "#/components/schemas/SearchTarget"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "$ref": "#/components/schemas/SearchSort"
            }
          },
          {
            "$ref": "#/components/parameters/OffsetQuery"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IllustPage"
                }
              }
            },
            "description": "Search results page."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "502": {
            "$ref": "#/components/responses/Upstream"
          }
        },
        "summary": "Search illustrations by keyword / tag.",
        "tags": [
          "search"
        ]
      }
    },
    "/search/users": {
      "get": {
        "operationId": "searchUsers",
        "parameters": [
          {
            "in": "query",
            "name": "word",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/OffsetQuery"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPreviewPage"
                }
              }
            },
            "description": "User search results."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "502": {
            "$ref": "#/components/responses/Upstream"
          }
        },
        "summary": "Search users by display name or account.",
        "tags": [
          "search"
        ]
      }
    },
    "/users/{id}/bookmarks": {
      "get": {
        "operationId": "listUserBookmarks",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdPath"
          },
          {
            "$ref": "#/components/parameters/RestrictQuery"
          },
          {
            "$ref": "#/components/parameters/MaxBookmarkIdQuery"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IllustPage"
                }
              }
            },
            "description": "Bookmarks page. `next_max_bookmark_id` is set when another page exists."
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "502": {
            "$ref": "#/components/responses/Upstream"
          }
        },
        "summary": "A user's bookmarked illustrations (cursor paged by `max_bookmark_id`).",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}/follow": {
      "delete": {
        "operationId": "deleteFollow",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdPath"
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "502": {
            "$ref": "#/components/responses/Upstream"
          }
        },
        "summary": "Unfollow a user.",
        "tags": [
          "users"
        ]
      },
      "put": {
        "operationId": "addFollow",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdPath"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FollowRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "502": {
            "$ref": "#/components/responses/Upstream"
          }
        },
        "summary": "Follow a user.",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}/following": {
      "get": {
        "operationId": "listUserFollowing",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdPath"
          },
          {
            "$ref": "#/components/parameters/RestrictQuery"
          },
          {
            "$ref": "#/components/parameters/OffsetQuery"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPreviewPage"
                }
              }
            },
            "description": "Following page."
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "502": {
            "$ref": "#/components/responses/Upstream"
          }
        },
        "summary": "Users that a user is following.",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}/illusts": {
      "get": {
        "operationId": "listUserIllusts",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdPath"
          },
          {
            "$ref": "#/components/parameters/IllustTypeQuery"
          },
          {
            "$ref": "#/components/parameters/OffsetQuery"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserIllustsPage"
                }
              }
            },
            "description": "User illusts page."
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "502": {
            "$ref": "#/components/responses/Upstream"
          }
        },
        "summary": "A user's own illustrations.",
        "tags": [
          "users"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "tags": [
    {
      "name": "system"
    },
    {
      "name": "auth"
    },
    {
      "name": "illusts"
    },
    {
      "name": "users"
    },
    {
      "name": "search"
    }
  ]
}